backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RAG-powered chatbot backend for Physical AI & Humanoid Robotics curriculum"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.0",
    "openai-agents>=0.6.2",
    "qdrant-client>=1.16.1",
    "trafilatura>=2.0.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]
